Add a callback parameter to '__connman_wispr_start'.
This callback provides two benefits.

The first is making static the implementation of the WISPr closure
function defined in src/service.c.

The second is that it will make easier closing several failure path
"holes" that currently do not properly "bookend" requests to
'__connman_wispr_start' by making the closure of those holes
independent of the function providing the closure.