Introduce the 'OnlineCheckIntervalStyle' setting.
This introduces the 'OnlineCheckIntervalStyle' setting which is the
style or mathematical series function used to compute the actual
time, in seconds, between two "ready" to "online" HTTP-based Internet
reachability checks. The value of which may be either "geometric" or
"fibonacci" with a default value of "geometric".

The "geometric" style (which is a formalization of the existing,
default series calculation) or function simply takes the square of the
online check interval. For example, at a check interval of 6, the
time, in seconds, is 36 (6^2) seconds.

The "fibonacci" style or function takes the value of the Fibonacci
sequence at the online check interval. For example, at a check interval
of 6, the time, in seconds, is 8 seconds.

The "fibonacci" series and style is more aggressive in check rate up to 12
steps (its equivalence point with "geometric" at 144 seconds) than
"geometric" but backs off far more aggressively past that point reaching
an hour at interval 19 which "geometric" does not reach until interval
60.

As the "EnableOnlineToReadyTransition" is refined and further developed,
the "fibonacci" style will allow for faster service failover and recovery
times for transient failures while backing off more aggressively for longer
duration outages where metered services like Cellular may be more sensitive
to a less aggresive backoff like "geometric".