service: Rename switch_default_service
The function 'switch_default_service' was formerly used in two contexts:

  1. From '__connman_service_move', which may be called from either the
     public D-Bus "MoveAfter" or "MoveBefore" method requests or from
     'connman_provider_set_split_routing' in src/provider.c.

  2. From 'service_update_preferred_order' in which the first argument
     is guaranteed to the default service.

In the D-Bus "MoveAfter" or "MoveBefore" context, the
'switch_default_service' is a misnomer since neither service argument,
in particular the first, is guaranteed to be the default service.
Consequently, 'switch_default_service' is renamed 'switch_service_order'
to reflect its varied usage in the above contexts.