service: Rename 'downgrade_state'.
This renames 'downgrade_state' to 'service_downgrade_online_state'.
This function does one thing and one thing only: downgrades the IPv4
and IPv6 configuration state from "online" to "ready" if and only if
those states are "online". The revised name more clearly indicates
that role and the action of the function and establishes a baseline
for a consistent nomenclature of related functions.