service: Rename 'apply_relevant_default_downgrade'.
This renames 'apply_relevant_default_downgrade' to
'service_downgrade_online_state_if_default'. This function does one
thing and one thing only: downgrades the IPv4 and IPv6 configuration
state from "online" to "ready" if and only if the specified service is
the default service (that is, has the default route) and its states are
"online". The revised name more clearly indicates that role and the
action of the function, is less ambiguous, and aligns with the newly-
renamed 'service_downgrade_online_state', which it, in turn, calls.