service: Add IP configuration type to 'cancel_online_check'.
WISPr / portal detection and HTTP-based "online" Internet reachability
checks are, by design, IP configuration-specific. There may be parallel,
independent requests or checks in flight: one for IPv4 and one for IPv6.
Consequently, when attempting to cancel such requests or checks, require
the IP configuration type to indicate which specific request or check is
to be canceled.