wispr: Add '__connman_wispr_cancel' interface.
This adds a heretofore-missing '__connman_wispr_cancel' interface. It
attempts to cancel a HTTP-based Internet reachability check for the
specified network service IP configuration type.

If a matching WISPr portal detect request or HTTP-based Internet
reachability check is found, the original callback specified with
'__connman_wispr_start' will be invoked with a success value of zero
('0') or false and an error value of -ECANCELED.