connection: Change the function signature of 'add_gateway'.
This changes the function signature of 'add_gateway' by migrating the
return type to a double-pointer parameter and changing the return type
to an 'int' for error status.

This allows passing the full range of failure errors to the caller and
allows the caller to pass those errors directly along as their own
return status.