connection: Rename 'choose_default_gateway'.
This renames 'choose_default_gateway' as 'yield_default_gateway'. The
latter more accurately reflects the action of the function which is
determining which of two gateway data should cede or yield the default
gateway and associated routes. The actual "choice" and acquisition of
the default gateway and associated routes happens outside of the
function.