Added @file comment to 'src/connman.h'.
While there is a very specific organization to the project headers and
symbols, it is not immediately evident at first glance to the uninitiated.

The addition of this file comment helps inform the uninitiated that
'connman.h' is reserved for private, cross-project interfaces which
differentiates it from those publicly-consumable headers in
'include/'.