service: Add '__connman_service_get_route_metric'.
This adds a new non-public service interface,
'__connman_service_get_route_metric', which attempts to get the route
metric/priority for the specified service based on the current service
and services state.

If the service is the default or if it is the only service, then the
metric is zero (0). Otherwise, a low-priority metric (metric > 0)
unique toservice and its underlying network interface is computed and
returned.