inet: Use '__connman_inet_rtnl_recv' for RTNL default route transaction.
This leverages '__connman_inet_rtnl_recv', introduced at commit
97967b4aae40 ("inet: Add '__connman_inet_rtnl_recv'."), to complete
the Routing Netlink (rtnl) default route addition or deletion
transaction in 'iproute_default_modify'.

Every rtnl route addition or deletion transaction consists of a
request and a response phase. The request phase contains the details
of the route to be added or deleted. The response phase contains the
status of whether the request succeeded. Unfortunately, to date,
'iproute_default_modify' did not include the response phase.
Therefore, it did not complete the route request transaction,
rendering the returns status incomplete in that it only reflected the
success or failure of the request phase.

Unfortunately, for callers of interfaces that, in turn, called
'iproute_default_modify', this made all routing requests appear to
artificially succeed even when, in fact, they were failing and elided
response status that callers need to conditionally act in the face of
the transaction failure.

By adding '__connman_inet_rtnl_recv' to complete the response phase of
the rtnl route transaction, the return status of
'iproute_default_modify', and those interfaces that call it, now
accurately reflect both the request and response status of the
transaction.