gateway: Handle -ESRCH in 'unset_default_gateway_route_common'.
This maps the error '-ESRCH' in 'unset_default_gateway_route_common'
to 0 ("Success").

Generally, we mandate that gateway routes follow the documented
lifecycle and finite state machine, using events and down- and upcalls
to drive the lifecycle.

There is one exception, however. When the Linux kernel
recognizes that the next hop (that is, the "via" or RTA_GATEWAY
portion of the route) for a route becomes unreachable, it is
automatically purged from the routing table with no
RTM_DELROUTE RTNL notification. Consequently, routes so purged
will return -ESRCH when we attempt to delete them here in the
mistaken belief they are still there.

By mapping -ESRCH to 0 ("Success") we ensure that gateway
configuration for such routes is not indefinitely stuck in the
"active" or "added" states but rather is correctly advanced to the
"removed" state.