main: Add comments tags for online check-related settings.
At this point, there are a sufficient number of settings in
'parse_config' that it is easy to lose track of where the block for
one settings handler ends and the next begins.

This adds comment tags for online check-related settings to make it
easy, at a glance, to see where the block for each is.