main: Add 'veth' to the default blacklist
Connman conflicts with Docker, breaking networking, as explained
and fixed in 9eb1772d31b6f ("Don't add route for invalid dst
and gateway address combinations"). However, that fix was reverted
in d8a634f20de3 due to conflicts with WireGuard.

This approach is simpler and does not conflict with the WireGuard
daemon, as tested in https://github.com/moby/moby/issues/49653