main: Introduce 'DefaultEnabledTechnologies' setting.
This introduces a new 'DefaultEnabledTechnologies' configuration
setting.

Since nearly the beginning of connman, it has behaved such that the
Ethernet technology is enabled (that is, powered) by
default. Currently, this is hard-coded behavior into
src/technology.c:technology_load.

However, if a different or additional technology is desired, there has
historically been no way to achieve this via configuration.

The 'DefaultEnabledTechnologies' key/value pair is a list of
technologies that are marked enabled (that is, powered) by default,
separated by commas ",". The default value for this entry when empty
is 'ethernet', matching the legacy behavior prior to the introduction
of this setting.